requirements.txt
# Core data
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=13.0.0

# ML
scikit-learn>=1.3.0
lightgbm>=4.1.0
xgboost>=2.0.0
catboost>=1.2.0
shap>=0.43.0

# Portfolio optimization
cvxpy>=1.4.0

# RL
stable-baselines3>=2.1.0
gymnasium>=0.29.0
shimmy>=1.3.0

# Data contracts
pydantic>=2.4.0

# Stats & quant
scipy>=1.11.0
statsmodels>=0.14.0
ta>=0.10.2

# Config
PyYAML>=6.0.1

# CLI
click>=8.1.7
tqdm>=4.66.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Experiment tracking
mlflow>=2.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0
